day2/lab 1/requirements.txt
langchain
langchain-openai
langchain-chroma
chromadb
pypdf
streamlit
python-dotenv
tiktoken
